[BUS-835] outsource isActive code into composable
