+++ components/Cms/Element/CmsElementCategoryNavigation.vue
@@ -11,12 +11,7 @@ const { navigationElements, loadNavigationElements } = useNavigation();
 
 await loadNavigationElements({ depth: 2 });
 
-const isActive = (path: Schemas['SeoUrl'][] | undefined, onlyExactMatch: boolean = false) => {
-    if (!path) return false;
-    const formattedPath = '/' + path[0]?.seoPathInfo;
-    const { path: currentPath } = useRoute();
-    return onlyExactMatch ? formattedPath === currentPath : currentPath.includes(formattedPath);
-};
+const { isActive } = useActivePath();
 </script>
 
 <template>
