+++ components/Navigation/NavigationLink.vue
@@ -26,14 +26,7 @@ if (isInternalLink) {
     await getInternalRoute();
 }
 
-const isActive = (path: Schemas['SeoUrl'][] | undefined, onlyExactMatch: boolean = false) => {
-    if (!path) return false;
-
-    const formattedPath = '/' + path[0]?.seoPathInfo;
-    const { path: currentPath } = useRoute();
-
-    return onlyExactMatch ? formattedPath === currentPath : currentPath.includes(formattedPath);
-};
+const { isActive } = useActivePath();
 </script>
 
 <template>
